+++ src/test/kotlin/me/bossm0n5t3r/leetcode/intersectionOfTwoLinkedLists/IntersectionOfTwoLinkedListsTest.kt
@@ -1,12 +1,12 @@
-package intersectionOfTwoLinkedLists
+package me.bossm0n5t3r.leetcode.intersectionOfTwoLinkedLists
 
+import me.bossm0n5t3r.leetcode.utils.ListNode
+import me.bossm0n5t3r.leetcode.utils.ListNodeUtil
 import org.junit.jupiter.api.Test
-import utils.ListNode
-import utils.ListNodeUtil
 import kotlin.test.assertEquals
 
 class IntersectionOfTwoLinkedListsTest {
-    private val intersectionOfTwoLinkedLists = IntersectionOfTwoLinkedLists()
+    private val intersectionOfTwoLinkedLists = IntersectionOfTwoLinkedLists.Solution()
     private val listNodeUtil = ListNodeUtil()
     private val basicSingleListNode = listNodeUtil.generateBasicSingleListNode()
